feat: Add temporary pokemon details activity
